Moved launch description also to utils
